+++ src/types/attachment.rs
@@ -18,7 +18,6 @@ fn bool_from_int_or_bool<'de, D: Deserializer<'de>>(d: D) -> Result<bool, D::Err
 #[non_exhaustive]
 pub struct Attachment {
     pub id: u64,
-    #[serde(default)]
     pub bug_id: u64,
     #[serde(default)]
     pub file_name: String,